+++ HISTORY.rst
@@ -12,6 +12,7 @@ Bug fixes:
 
 - ``Date`` fields correctly deserializes to a ``datetime.date`` object when ``python-dateutil`` is not installed. Thanks @malexer for the catch and patch.
 - Fix bug that raised an ``AttributeError`` when using a class-based validator.
+- Fix `as_string` behavior of Number fields when serializing to default value.
 
 Support: